+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - New flag `--hash` which allow to specify the hash algorithm used for computing blocks checksums.
 
+- Server and client check that they are using the same version of the procotol before going on.
+
 # 1.0.0 (2024-02-01)
 
 Initial release